Fixed Makefile to use gnu-efi installed location in /usr/lib and /usr/include
Note: According #4, the make process uses gnu-efi, which uses standard
locations, which are also the same as what are used in Debian, and
should be the same in most linux distributions. Probably not NixOS, but
that's Nix.

Ideally, the best option would be to use Docker to build the EFI, then
literally any machine that can run Docker will be able to build this
without requiring installation of even build-essentials or gnu-efi,
which would also allow automated building.